feat: area de detalhes do pokemon
