+++ package.json
@@ -19,6 +19,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.8",
+    "react-router-dom": "^6.23.0",
     "redux": "^5.0.1"
   },
   "devDependencies": {
