+++ src/main.tsx
@@ -4,12 +4,15 @@ import './index.css'
 import App from './App'
 import { Provider } from 'react-redux'
 import store from './redux/store'
+import { BrowserRouter } from 'react-router-dom'
 
 ReactDOM.render(
   <React.StrictMode>
-    <Provider store={store}>
-      <App />
-    </Provider>
+    <BrowserRouter>
+      <Provider store={store}>
+        <App />
+      </Provider>
+    </BrowserRouter>
   </React.StrictMode>,
   document.getElementById('root')
 )
